clean up slack alerting
